A lot of changes were made:
-StackToken class was made. This is nothing but a stack of tokens. The
only place this is used is for the operatorStack in the shuntingYard
algorithm. This is essentially useless.

-Precedence added to the TokenValue class, precedence of +- is 1, */ is
2, ^ is 3. This isn't really required to implement the RPN. The
precedence is set using the setPrecedence() present in the resources
class.

-Added a new constructor to the TokenValue class in order to make
storing precedence possible without making any change where the earlier
constructor was used. Done so that the code looks class-ier.

-Translated the ShuntingYard algorithm Psuedocode to Code, works
flawlessly. This function returns a TokenList which has to be processed
by the RPN.